General Update
Deleted files that were not going to be used, added bushes and trees to the main gameplay map, and made the firebolt ability apply damage to a character, as well as cost mana, and implemented a mana regeneration for the player.

I also changed the behavior of the firebolt spawning. Instead of spawning attached to the character's socket, and then detaching in the notify, it now spawns in the notify and a force is immediately applied to it