Improved for Replication
Many changes happened in order to make the elements of the game replicate in multiplayer mode. Currently, the character is on a process of change to make the GAS replicate perfectly, and the HUD will later be improved so that the life is managed by a delegate.  But it's all still a work in process